requirements.txt
requests==2.31.0
web3==6.15.1
python-dotenv==1.0.1
loguru==0.7.2
SQLAlchemy==2.0.28
aiosqlite==0.20.0
greenlet==3.0.3
alembic==1.13.1
web3==6.15.1
fake-useragent==1.5.0
aiohttp==3.9.3
aiohttp-socks==0.8.4
curl_cffi==0.6.2
art~=6.2